# swagger.json
{
          "openapi": "3.1.0",
          "info": {
                    "title": "FastAPI Boilerplate",
                    "description": "FastAPI Boilerplate",
                    "version": "1.0.0"
          },
          "paths": {
...
                    "/api/v1/roles": {
                              "get": {
                                        "tags": [
                                                  "Roles"
                                        ],
                                        "summary": "Get User Roles",
                                        "description": "Get all user roles",
                                        "operationId": "get_user_roles_api_v1_roles_get",
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  }
                                        },
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ]
                              },
                              "post": {
                                        "tags": [
                                                  "Roles"
                                        ],
                                        "summary": "Create User Role",
                                        "operationId": "create_user_role_api_v1_roles_post",
                                        "requestBody": {
                                                  "content": {
                                                            "application/json": {
                                                                      "schema": {
                                                                                "$ref": "#/components/schemas/CreateUserRoleRequest"
                                                                      }
                                                            }
                                                  },
                                                  "required": true
                                        },
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        },
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ]
                              }
                    },
                    "/api/v1/roles/{user_role_id}": {
                              "get": {
                                        "tags": [
                                                  "Roles"
                                        ],
                                        "summary": "Get User Role",
                                        "operationId": "get_user_role_api_v1_roles__user_role_id__get",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "user_role_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "User Role Id"
                                                            }
                                                  }
                                        ],
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              },
                              "put": {
                                        "tags": [
                                                  "Roles"
                                        ],
                                        "summary": "Update User Role",
                                        "operationId": "update_user_role_api_v1_roles__user_role_id__put",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "user_role_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "User Role Id"
                                                            }
                                                  }
                                        ],
                                        "requestBody": {
                                                  "required": true,
                                                  "content": {
                                                            "application/json": {
                                                                      "schema": {
                                                                                "$ref": "#/components/schemas/CreateUserRoleRequest"
                                                                      }
                                                            }
                                                  }
                                        },
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              },
                              "delete": {
                                        "tags": [
                                                  "Roles"
                                        ],
                                        "summary": "Delete User Role",
                                        "operationId": "delete_user_role_api_v1_roles__user_role_id__delete",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "user_role_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "User Role Id"
                                                            }
                                                  }
                                        ],
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              }
                    },
                    "/api/v1/roles/{user_role_id}/permissions": {
                              "get": {
                                        "tags": [
                                                  "Roles"
                                        ],
                                        "summary": "Get User Role Permissions",
                                        "operationId": "get_user_role_permissions_api_v1_roles__user_role_id__permissions_get",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "user_role_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "User Role Id"
                                                            }
                                                  }
                                        ],
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              }
                    },
...
                    "/api/v1/users/{user_id}/update/role/{role_id}": {
                              "patch": {
                                        "tags": [
                                                  "Roles"
                                        ],
                                        "summary": "Update User Role",
                                        "operationId": "update_user_role_api_v1_users__user_id__update_role__role_id__patch",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "user_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "User Id"
                                                            }
                                                  },
                                                  {
                                                            "name": "role_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "Role Id"
                                                            }
                                                  }
                                        ],
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              }
                    },
...
          },
          "components": {
                    "schemas": {
...
                              "CreateUserRoleRequest": {
                                        "properties": {
                                                  "role_name": {
                                                            "type": "string",
                                                            "title": "Role Name",
                                                            "description": "name of the role",
                                                            "examples": [
                                                                      "admin",
                                                                      "user",
                                                                      "moderator"
                                                            ]
                                                  }
                                        },
                                        "type": "object",
                                        "required": [
                                                  "role_name"
                                        ],
                                        "title": "CreateUserRoleRequest"
                              },
                              "HTTPValidationError": {
                                        "properties": {
                                                  "detail": {
                                                            "items": {
                                                                      "$ref": "#/components/schemas/ValidationError"
                                                            },
                                                            "type": "array",
                                                            "title": "Detail"
                                                  }
                                        },
                                        "type": "object",
                                        "title": "HTTPValidationError"
                              },
...
                              "ValidationError": {
                                        "properties": {
                                                  "loc": {
                                                            "items": {
                                                                      "anyOf": [
                                                                                {
                                                                                          "type": "string"
                                                                                },
                                                                                {
                                                                                          "type": "integer"
                                                                                }
                                                                      ]
                                                            },
                                                            "type": "array",
                                                            "title": "Location"
                                                  },
                                                  "msg": {
                                                            "type": "string",
                                                            "title": "Message"
                                                  },
                                                  "type": {
                                                            "type": "string",
                                                            "title": "Error Type"
                                                  }
                                        },
                                        "type": "object",
                                        "required": [
                                                  "loc",
                                                  "msg",
                                                  "type"
                                        ],
                                        "title": "ValidationError"
                              }
                    },
                    "securitySchemes": {
                              "HTTPBearer": {
                                        "type": "http",
                                        "scheme": "bearer"
                              }
                    }
          }
}